Moved tests to DeployMojoTest.
